frontend/requirements.txt
supabase 
streamlit
streamlit_option_menu
python-dotenv